refactor(home-planner): HomePlanner 모델을 widgets 레이어로 이동 (#188)
