+++ src/features/home/index.ts
@@ -11,12 +11,7 @@ export type {
   TaskSplitGroup,
   TaskSplitItem,
 } from "./task-basket";
-export { useHomePlanner } from "./model/useHomePlanner";
-export { useHomePlannerCalendar } from "./model/useHomePlannerCalendar";
-export { useHomePlannerQueries } from "./model/useHomePlannerQueries";
-export { useMergedTasks } from "./model/useMergedTasks";
 export { useDayPlanId } from "./model/useDayPlanId";
-export { usePlannerStatus } from "./model/usePlannerStatus";
 export { useAiArrangeNoticeStore } from "./model/aiArrangeNotice.store";
 export {
   useAiArrangeScheduleMutation,
@@ -27,4 +22,13 @@ export {
   type UpdateScheduleStatusPayload,
 } from "./model/useScheduleMutations";
 
-export { DAYS_IN_WEEK, END_HOUR, START_HOUR, WEEKDAY_LABELS, isSameDate } from "./model/calendar";
+export {
+  DAYS_IN_WEEK,
+  END_HOUR,
+  START_HOUR,
+  WEEKDAY_LABELS,
+  isSameDate,
+  addDays,
+  toStartOfWeek,
+  getRepresentativeMonthIndex,
+} from "./model/calendar";
