+++ src/widgets/home-planner/index.ts
@@ -1,1 +1,6 @@
 export { HomePlanner } from "./ui/HomePlanner";
+export { useHomePlanner } from "./model/useHomePlanner";
+export { useHomePlannerCalendar } from "./model/useHomePlannerCalendar";
+export { useHomePlannerQueries } from "./model/useHomePlannerQueries";
+export { useMergedTasks } from "./model/useMergedTasks";
+export { usePlannerStatus } from "./model/usePlannerStatus";
