+++ src/widgets/home-planner/model/date.ts
@@ -1,4 +1,4 @@
-import { KOREAN_WEEKDAY_LABELS_BY_GETDAY } from "./calendar";
+import { WEEKDAY_LABELS } from "@/features/home";
 
 export function formatDateToYmd(date: Date) {
   const year = date.getFullYear();
@@ -11,6 +11,6 @@ export function formatScheduleLabel(date: Date) {
   if (Number.isNaN(date.getTime())) return "";
   const month = String(date.getMonth() + 1).padStart(2, "0");
   const day = String(date.getDate()).padStart(2, "0");
-  const weekday = KOREAN_WEEKDAY_LABELS_BY_GETDAY[date.getDay()] ?? "";
+  const weekday = WEEKDAY_LABELS[date.getDay()] ?? "";
   return `${month}.${day} (${weekday}) 일정`;
 }
