+++ src/widgets/home-planner/model/planPresenceViewModel.ts
